Add files via upload
STEP 1.
one.cpp is my first step. It is a follow-up workout on a YouTube video by Derek Banas. i am attaching the link below. 
https://www.youtube.com/watch?v=Rub-JsjMhWY
This video is very straight forward and informative. 

STEP 2.
As the next step I started solving problems in HackerRank. 

So far this is all that I have done!